+++ src/components/index.js
@@ -1,5 +1,6 @@
 import App from './App/index';
 import Example from './Example/index';
+import FmaMap from './FmaMap';
 import NotFoundPage from './NotFoundPage/index';
 import Root from './Root';
 import TestComponent from './TestComponent';
@@ -7,6 +8,7 @@ import TestComponent from './TestComponent';
 export {
   App,
   Example,
+  FmaMap,
   NotFoundPage,
   Root,
   TestComponent,
